Fixed TestCluster for the average traj. activation
Originally, the activation from the trajectory was computed by adding up the activation from each of the points and capping the sum at 1.  It was changed to be computed by taking the *average* activation from the points in the trajectory.  This change has been in place for a while, but the TestCluster file was not up to date on this, so it was returning failures.  That test file is now updated.